Waypoint generator tests (#198)
test_waypoint_generator and various cleanup

+++ local_planner/src/nodes/local_planner_node.cpp
@@ -144,19 +144,19 @@ void LocalPlannerNode::readParams() {
 
   // Read in parameter for waypoint generator
   waypointGenerator_params new_params;
-  nh_.param<double>("goal_acceptance_radius_in",
-                    new_params.goal_acceptance_radius_in, 0.5);
-  nh_.param<double>("goal_acceptance_radius_out",
-                    new_params.goal_acceptance_radius_out, 1.5);
-  nh_.param<double>("factor_close_to_goal_start_speed_limitation",
-                    new_params.factor_close_to_goal_start_speed_limitation,
-                    3.0);
-  nh_.param<double>("factor_close_to_goal_stop_speed_limitation",
-                    new_params.factor_close_to_goal_stop_speed_limitation, 4.0);
-  nh_.param<double>("max_speed_close_to_goal_factor",
-                    new_params.max_speed_close_to_goal_factor, 0.1);
-  nh_.param<double>("min_speed_close_to_goal",
-                    new_params.min_speed_close_to_goal, 0.5);
+  nh_.param<float>("goal_acceptance_radius_in",
+                   new_params.goal_acceptance_radius_in, 0.5f);
+  nh_.param<float>("goal_acceptance_radius_out",
+                   new_params.goal_acceptance_radius_out, 1.5f);
+  nh_.param<float>("factor_close_to_goal_start_speed_limitation",
+                   new_params.factor_close_to_goal_start_speed_limitation,
+                   3.0f);
+  nh_.param<float>("factor_close_to_goal_stop_speed_limitation",
+                   new_params.factor_close_to_goal_stop_speed_limitation, 4.0f);
+  nh_.param<float>("max_speed_close_to_goal_factor",
+                   new_params.max_speed_close_to_goal_factor, 0.1f);
+  nh_.param<float>("min_speed_close_to_goal",
+                   new_params.min_speed_close_to_goal, 0.5f);
 
   wp_generator_->param_ = new_params;
 }
